requirements.txt
pyspark==3.5.0
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
xgboost==2.0.0
lightgbm==4.1.0
matplotlib==3.8.0
seaborn==0.13.0
jupyter==1.0.0
pyyaml==6.0.1
pytest==7.4.0
black==23.9.0
pylint==2.17.0
